Remove legacy code, update algorithms and signers/verifiers
